Qgs3DMapCanvas + Qgs3DWindow = <3
