+++ python/PyQt6/3d/3d_auto.sip
@@ -1,10 +1,11 @@
 // Include auto-generated SIP files
 %Include auto_generated/processing/qgs3dalgorithms.sip
 %Include auto_generated/qgs3d.sip
+%Include auto_generated/qgs3dmapcanvas.sip
 %Include auto_generated/qgs3dmapscene.sip
 %Include auto_generated/qgs3dmapsettings.sip
+%Include auto_generated/qgs3dmaptool.sip
 %Include auto_generated/qgs3dtypes.sip
-%Include auto_generated/qgs3dwindow.sip
 %Include auto_generated/qgsabstractvectorlayer3drenderer.sip
 %Include auto_generated/qgscameracontroller.sip
 %Include auto_generated/qgscamerapose.sip
