+++ python/PyQt6/3d/auto_generated/qgs3dmaptool.sip.in
@@ -1,0 +1,64 @@
+/************************************************************************
+ * This file has been generated automatically from                      *
+ *                                                                      *
+ * src/3d/qgs3dmaptool.h                                                *
+ *                                                                      *
+ * Do not edit manually ! Edit header and run scripts/sipify.pl again   *
+ ************************************************************************/
+
+
+
+
+
+class Qgs3DMapTool : QObject
+{
+%Docstring(signature="appended")
+Base class for map tools operating on 3D map canvas.
+%End
+
+%TypeHeaderCode
+#include "qgs3dmaptool.h"
+%End
+  public:
+    Qgs3DMapTool( Qgs3DMapCanvas *canvas );
+
+    virtual void mousePressEvent( QMouseEvent *event );
+    virtual void mouseReleaseEvent( QMouseEvent *event );
+    virtual void mouseMoveEvent( QMouseEvent *event );
+    virtual void keyPressEvent( QKeyEvent *event );
+
+    virtual void activate();
+%Docstring
+Called when set as currently active map tool
+%End
+
+    virtual void deactivate();
+%Docstring
+Called when map tool is being deactivated
+%End
+
+    virtual QCursor cursor() const;
+%Docstring
+Mouse cursor to be used when the tool is active
+%End
+
+    virtual bool allowsCameraControls() const;
+%Docstring
+Whether the default mouse controls to zoom/pan/rotate camera can stay enabled
+while the tool is active. This may be useful for some basic tools using just
+mouse clicks (e.g. identify, measure), but it could be creating conflicts when used
+with more advanced tools. Default implementation returns ``True``.
+%End
+
+    Qgs3DMapCanvas *canvas();
+
+  protected:
+};
+
+/************************************************************************
+ * This file has been generated automatically from                      *
+ *                                                                      *
+ * src/3d/qgs3dmaptool.h                                                *
+ *                                                                      *
+ * Do not edit manually ! Edit header and run scripts/sipify.pl again   *
+ ************************************************************************/
