+++ src/3d/CMakeLists.txt
@@ -11,13 +11,12 @@ set(QGIS_3D_SRCS
   qgs3danimationsettings.cpp
   qgs3dexportobject.cpp
   qgs3dmapexportsettings.cpp
-  qgs3dmapcanvas.cpp
   qgs3dmapscene.cpp
   qgs3dmapsettings.cpp
   qgs3dmaptool.cpp
+  qgs3dmapcanvas.cpp
   qgs3dsceneexporter.cpp
   qgs3dutils.cpp
-  qgs3dwindow.cpp
   qgscameracontroller.cpp
   qgscamerapose.cpp
   qgsfeature3dhandler_p.cpp
@@ -113,14 +112,13 @@ set(QGIS_3D_HDRS
   qgs3danimationsettings.h
   qgs3daxis.h
   qgs3daxissettings.h
-  qgs3dmapcanvas.h
   qgs3dmapscene.h
   qgs3dmapsettings.h
   qgs3dmaptool.h
   qgs3dsceneexporter.h
   qgs3dtypes.h
   qgs3dutils.h
-  qgs3dwindow.h
+  qgs3dmapcanvas.h
   qgsaabb.h
   qgsabstract3dengine.h
   qgsabstractvectorlayer3drenderer.h
