+++ src/3d/qgs3dmapcanvas.cpp
@@ -13,24 +13,60 @@
  *                                                                         *
  ***************************************************************************/
 
+#include <Qt3DCore/QAspectEngine>
+#include <Qt3DCore/QEntity>
+#if QT_VERSION >= QT_VERSION_CHECK(6, 0, 0)
+#include <Qt3DCore/QCoreAspect>
+#endif
+#include <Qt3DExtras/QForwardRenderer>
+#include <Qt3DRender/QRenderSettings>
+#include <Qt3DRender/QRenderAspect>
+#include <Qt3DInput/QInputAspect>
+#include <Qt3DInput/QInputSettings>
+#include <Qt3DLogic/QLogicAspect>
+#include <Qt3DRender/QCamera>
+
 #include "qgs3dmapcanvas.h"
 
-#include <QBoxLayout>
-#include <Qt3DRender/QRenderCapture>
 #include <Qt3DLogic/QFrameAction>
-#include <QMouseEvent>
-
-#include "qgscameracontroller.h"
-#include "qgs3dmapsettings.h"
 #include "qgs3dmapscene.h"
-#include "qgs3dmaptool.h"
 #include "qgswindow3dengine.h"
-#include "qgssettings.h"
+#include "qgs3dmapsettings.h"
+#include "qgs3dmaptool.h"
 #include "qgstemporalcontroller.h"
 
-Qgs3DMapCanvas::Qgs3DMapCanvas( QWidget *parent )
-  : QWidget( parent )
+Qgs3DMapCanvas::Qgs3DMapCanvas()
+  : m_aspectEngine( new Qt3DCore::QAspectEngine )
+  , m_renderAspect( new Qt3DRender::QRenderAspect )
+  , m_inputAspect( new Qt3DInput::QInputAspect )
+  , m_logicAspect( new Qt3DLogic::QLogicAspect )
+  , m_renderSettings( new Qt3DRender::QRenderSettings )
+  , m_forwardRenderer( new Qt3DExtras::QForwardRenderer )
+  , m_defaultCamera( new Qt3DRender::QCamera )
+  , m_inputSettings( new Qt3DInput::QInputSettings )
+  , m_root( new Qt3DCore::QEntity )
+  , m_userRoot( nullptr )
+  , m_initialized( false )
 {
+  setSurfaceType( QSurface::OpenGLSurface );
+
+  // register aspects
+#if QT_VERSION >= QT_VERSION_CHECK(6, 0, 0)
+  m_aspectEngine->registerAspect( new Qt3DCore::QCoreAspect );
+#endif
+  m_aspectEngine->registerAspect( m_renderAspect );
+  m_aspectEngine->registerAspect( m_inputAspect );
+  m_aspectEngine->registerAspect( m_logicAspect );
+
+  m_defaultCamera->setParent( m_root );
+  m_forwardRenderer->setCamera( m_defaultCamera );
+  m_forwardRenderer->setSurface( this );
+  m_renderSettings->setActiveFrameGraph( m_forwardRenderer );
+  m_inputSettings->setEventSource( this );
+
+
+
+
   const QgsSettings setting;
   mEngine = new QgsWindow3DEngine( this );
 
@@ -41,8 +77,8 @@ Qgs3DMapCanvas::Qgs3DMapCanvas( QWidget *parent )
     emit savedAsImage( mCaptureFileName );
   } );
 
-  mEngine->window()->setCursor( Qt::OpenHandCursor );
-  mEngine->window()->installEventFilter( this );
+  setCursor( Qt::OpenHandCursor );
+  installEventFilter( this );
 }
 
 Qgs3DMapCanvas::~Qgs3DMapCanvas()
@@ -54,17 +90,76 @@ Qgs3DMapCanvas::~Qgs3DMapCanvas()
   mScene = nullptr;
   mMap->deleteLater();
   mMap = nullptr;
+
+
+  delete m_aspectEngine;
 }
 
-void Qgs3DMapCanvas::resizeEvent( QResizeEvent *ev )
+void Qgs3DMapCanvas::setRootEntity( Qt3DCore::QEntity *root )
 {
-  QWidget::resizeEvent( ev );
+  if ( m_userRoot != root )
+  {
+    if ( m_userRoot != nullptr )
+      m_userRoot->setParent( static_cast<Qt3DCore::QNode *>( nullptr ) );
+    if ( root != nullptr )
+      root->setParent( m_root );
+    m_userRoot = root;
+  }
+}
 
-  if ( !mScene )
-    return;
+void Qgs3DMapCanvas::setActiveFrameGraph( Qt3DRender::QFrameGraphNode *activeFrameGraph )
+{
+  m_renderSettings->setActiveFrameGraph( activeFrameGraph );
+}
+
+Qt3DRender::QFrameGraphNode *Qgs3DMapCanvas::activeFrameGraph() const
+{
+  return m_renderSettings->activeFrameGraph();
+}
+
+Qt3DExtras::QForwardRenderer *Qgs3DMapCanvas::defaultFrameGraph() const
+{
+  return m_forwardRenderer;
+}
+
+Qt3DRender::QCamera *Qgs3DMapCanvas::camera() const
+{
+  return m_defaultCamera;
+}
+
+Qt3DRender::QRenderSettings *Qgs3DMapCanvas::renderSettings() const
+{
+  return m_renderSettings;
+}
+
+void Qgs3DMapCanvas::showEvent( QShowEvent *e )
+{
+  if ( !m_initialized )
+  {
+    m_root->addComponent( m_renderSettings );
+    m_root->addComponent( m_inputSettings );
+    m_aspectEngine->setRootEntity( Qt3DCore::QEntityPtr( m_root ) );
+
+    m_initialized = true;
+  }
+  QWindow::showEvent( e );
+}
+
+void Qgs3DMapCanvas::resizeEvent( QResizeEvent * )
+{
+  m_defaultCamera->setAspectRatio( float( width() ) / std::max( 1.f, static_cast<float>( height() ) ) );
 
   mEngine->setSize( size() );
 }
+
+
+
+
+
+
+
+
+
 
 void Qgs3DMapCanvas::setMap( Qgs3DMapSettings *map )
 {
@@ -177,7 +272,7 @@ void Qgs3DMapCanvas::setMapTool( Qgs3DMapTool *tool )
   if ( mMapTool && !tool )
   {
     mScene->cameraController()->setEnabled( true );
-    mEngine->window()->setCursor( Qt::OpenHandCursor );
+    setCursor( Qt::OpenHandCursor );
   }
   else if ( !mMapTool && tool )
   {
@@ -192,14 +287,14 @@ void Qgs3DMapCanvas::setMapTool( Qgs3DMapTool *tool )
   if ( mMapTool )
   {
     mMapTool->activate();
-    mEngine->window()->setCursor( mMapTool->cursor() );
+    setCursor( mMapTool->cursor() );
   }
 
 }
 
 bool Qgs3DMapCanvas::eventFilter( QObject *watched, QEvent *event )
 {
-  if ( watched != mEngine->window() )
+  if ( watched != this )
     return false;
 
   if ( event->type() == QEvent::ShortcutOverride )
