+++ src/app/3d/qgs3dnavigationwidget.cpp
@@ -171,23 +171,3 @@ void Qgs3DNavigationWidget::updateFromCamera()
   mCameraInfoItemModel->setData( mCameraInfoItemModel->index( 6, 1 ), QStringLiteral( "%1" ).arg( m3DMapCanvas->cameraController()->lookingAtPoint().y() ) );
   mCameraInfoItemModel->setData( mCameraInfoItemModel->index( 7, 1 ), QStringLiteral( "%1" ).arg( m3DMapCanvas->cameraController()->lookingAtPoint().z() ) );
 }
-
-void Qgs3DNavigationWidget::resizeEvent( QResizeEvent *ev )
-{
-  QWidget::resizeEvent( ev );
-
-  QSize size = ev->size();
-  emit sizeChanged( size );
-}
-
-void Qgs3DNavigationWidget::hideEvent( QHideEvent *ev )
-{
-  QWidget::hideEvent( ev );
-  emit sizeChanged( QSize( 0, 0 ) );
-}
-
-void Qgs3DNavigationWidget::showEvent( QShowEvent *ev )
-{
-  QWidget::showEvent( ev );
-  emit sizeChanged( size() );
-}
